Removed old job page
