--- src/app/components/headers/job-actions-header/job-actions-header.component.scss
@@ -1,14 +1,0 @@
-@import '../../../styles/colors';
-@import '../../../styles/dimensions';
-@import '../../../styles/typography';
-
-.job-actions-header {
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  display: flex;
-  background: $white url('/assets/images/job-header-background.jpg') no-repeat right top !important;
-  background-size: cover !important;
-  color: white;
-  min-height: 300px;
-}
