+++ src/app/components/pages/job-page/job-page.component.scss
@@ -1,7 +1,6 @@
 @import '../../../styles/colors';
 @import '../../../styles/dimensions';
 
-
 .job-page-grid {
   @media screen and (min-width: $tablet-min-width) {
     padding: 0 50px 50px 50px;
