+++ src/app/components/sections/job-banner-section/job-banner-section.component.scss
@@ -2,7 +2,7 @@
 @import '../../../styles/dimensions';
 @import '../../../styles/typography';
 
-.header-container {
+.banner-container {
   background: $white url('/assets/images/job-background-computer.jpg') no-repeat center bottom !important;
   background-size: cover !important;
   min-height: 300px;
