--- src/app/components/sections/job-description-section2/job-description-section2.component.ts
@@ -1,31 +1,0 @@
-import {Component} from '@angular/core';
-import {Input} from '@angular/core';
-import {Job} from '../../../models/api-models/job/job'
-import {BaseComponent} from '../../base.component';
-import {SystemLanguagesResolver} from '../../../resolvers/system-languages/system-languages.resolver';
-import {UserResolver} from '../../../resolvers/user/user.resolver';
-
-@Component({
-  selector: 'job-description-section2',
-  template: `
-    <basic-title-text
-      [text]="'job.description.section.title' | translate "
-      [uppercase]="true"
-      color="pink"
-      fontSize="medium">
-    </basic-title-text>
-    <basic-text
-      [unsafeHtml]="'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.<br><br> Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.<br><br> Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.<br><br> Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.' || job.translatedText.descriptionHtml"
-      color="gray">
-    </basic-text>`
-})
-export class JobDescriptionSection2Component extends BaseComponent {
-  @Input() job = null as Job;
-
-  public constructor (
-    protected systemLanguagesResolver: SystemLanguagesResolver,
-    protected userResolver: UserResolver,
-  ) {
-    super(systemLanguagesResolver, userResolver);
-  }
-}
