+++ src/assets/i18n/en.json
@@ -1,16 +1,4 @@
 {
-  "job.actions.section.read.more": "Read more about",
-  "job.recruiter.section.title": "Responsible recruiter",
-  "job.important.information.section.hours": "{{hours}} hours",
-  "job.important.information.section.salary.per.hour": "{{grossSalaryPerHour}} gross salary",
-  "job.important.information.section.total.salary": "{{totalGrossSalary}} total salary ({{totalNetSalary}} after tax)",
-  "job.skills.section.title": "You have",
-  "job.interests.section.title": "You are",
-  "job.tasks.section.title": "Tasks",
-  "job.important.information.section.title": "Important information",
-  "job.description.section.title": "About the job",
-  "job.short.description.section.title": "Short about the job",
-  "job.company.short.description.section.title": "Short about {{company}}",
   "already.registered.modal.button": "Log in",
   "already.registered.modal.info": "It seems that there already is an account using {{emailOrPhone}}, so we suggest that you log in instead. If you forgot your password you can reset your password by clicking \"Forgotten your password?\" above",
   "already.registered.modal.title": "{{emailOrPhone}} is already registered",
@@ -259,6 +247,7 @@
   "job.actions.section.login": "Log in to apply",
   "job.actions.section.offer": "You have received an offer to do this job. Answer within {{hours}} hours and {{minutes}} minutes.",
   "job.actions.section.performed": "Job performed",
+  "job.actions.section.read.more": "Read more about",
   "job.actions.section.register": "Create an account",
   "job.additional.user.info.modal.button": "Continue",
   "job.additional.user.info.modal.title": "To maximize your chances to get this job please consider providing the following information",
@@ -270,14 +259,25 @@
   "job.card.temporary": "Temporary",
   "job.card.until.further.notice": "Indefinite duration",
   "job.comments.section.title": "Comments",
+  "job.company.short.description.section.title": "Short about {{company}}",
   "job.description.section.header": "Description",
+  "job.description.section.title": "About the job",
+  "job.important.information.section.hours": "{{hours}} hours",
+  "job.important.information.section.salary.per.hour": "{{grossSalaryPerHour}} gross salary",
+  "job.important.information.section.title": "Important information",
+  "job.important.information.section.total.salary": "{{totalGrossSalary}} total salary ({{totalNetSalary}} after tax)",
   "job.information.section.header": "Information",
   "job.location.section.header": "Location",
+  "job.recruiter.section.title": "Responsible recruiter",
   "job.scope.section.grossSalary": "{{grossSalary}} gross salary (before taxes)",
   "job.scope.section.header": "Scope",
   "job.scope.section.hours": "{{hours}} hours",
   "job.scope.section.netSalary": "{{netSalary}} net salary (after taxes)",
   "job.scope.section.until.further.notice": "Until further notice",
+  "job.short.description.section.title": "Short about the job",
+  "job.tasks.section.title": "Tasks",
+  "job.you.are.section.title": "You are",
+  "job.you.have.section.title": "You have",
   "jobs.title": "All jobs ({{nbrOfJobs}})",
   "language.proficiency.basic": "Basic knowledge",
   "language.proficiency.beginner": "Beginner",
